[REF] Refactor hr_biometric_machine
