--- __init__.py
@@ -1,3 +1,0 @@
-import biometric_machine
-import report
-import wizard
