+++ hr_biometric_machine/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+import biometric_machine
+import report
+import wizard
