+++ hr_biometric_machine/__openerp__.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 {
     "name" : "Biometric Device Integration",
     "version" : "1.0",
@@ -8,9 +10,9 @@
     "depends" : ["base","hr"],
     "init_xml" : [],
     "data" : [
-        "biometric_machine_view.xml",
+        "views/biometric_machine_view.xml",
         "report/daily_attendance_view.xml",
-        "schedule.xml",
+        "views/schedule.xml",
         "wizard/schedule_wizard.xml",
     ],
     "active": False,
