+++ hr_biometric_machine/report/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+import daily_attendance
