+++ hr_biometric_machine/report/daily_attendance_view.xml
