+++ hr_biometric_machine/static/description/index.html
