+++ hr_biometric_machine/views/biometric_machine_view.xml
