+++ hr_biometric_machine/views/schedule.xml
