+++ hr_biometric_machine/wizard/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+import schedule_wizard
