+++ hr_biometric_machine/wizard/schedule_wizard.py
