--- wizard/__init__.py
@@ -1,1 +1,0 @@
-import schedule_wizard
